requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
pydantic
redis
python-dotenv
